requirements.txt

flask>=2.3.0
tensorflow>=2.13.0
pillow>=10.0.0
numpy>=1.24.0
opencv-python>=4.8.0
gunicorn>=21.2.0
requests>=2.31.0
matplotlib>=3.7.0
seaborn>=0.12.3
scikit-learn>=1.3.0
scipy>=1.11.0
tqdm>=4.65.0
# imutils removed – brain-contour cropping now uses pure OpenCV (no extra dep)